Refactoring file structure
Ran into issues during docker deployment with local imports.  Docker container now running without crashing
