+++ ap002-bloombreaks-app/src/main/python/application/consumer/core/facade/Base.py
