+++ ap002-bloombreaks-app/src/main/python/application/consumer/core/facade/model/infrastructure/configuration/config.json
