+++ ap002-bloombreaks-app/src/main/python/application/consumer/core/facade/model/infrastructure/repository/AuthDB.py
@@ -1,4 +1,4 @@
-from main.python.infrastructure.repository.DB import PostgresDB
+from DB import PostgresDB
 
 class authDB(PostgresDB):
     def __init__(self, timeout=30, max_retries=5):
